requirements.txt
requests==2.31.0
pymongo==4.6.1
openai==1.56.1
langchain==0.1.9
langchain-openai==0.0.8
streamlit==1.32.0
python-dotenv==1.0.1
newsapi-python==0.2.7 